fix(scripts/commit): strip rogue commit arg coming from yarn commit
For some reason invoking this script with `yarn commit` now passes through `'commit'` in the arguments that get parsed as "raw git arguments" in the adapter which blows up said parsing behavior. I'm not sure if this was related to a yarn change (`yarn [script] args` vs. `yarn [script] -- args`) or something else...

Either way, this is a band-aid to filter out the rogue argument for now.