feat: 🔪 remove build script and related configuration
We don't use any of the build-related tooling that this library exports. This
removes the script and all related configuration and dependencies and replaces
the build process with a simple Babel CLI based build.

BREAKING CHANGE: This removes the `hover-scripts build` script, the exported
Babel configuration, and the exported Rollup configuration. It also removes the
built-in Babel integration in the Jest configuation.