build: move shared TypeScript configuration to project root instead of extending
For consistency, up until now, the shared TypeScript configuration has lived in
`src/config` with all of the other shared configurations that this package
provides. For all of the other configurations we're able to simple re-export
the configuration at the project root to provide it to consumers at
`require('@hover/javascript/[config]')`. Because the TypeScript configuration
is JSON, however, we were using `extends` instead of re-exporting. This has been
causing all kinds of trouble within this package since we were extending from the
`dist` directory. This finally moves the canonical TypeScript configuration
to the root, so that we don't need to `extends` from `dist/` while preserving
`src/config/tsconfig.json` via `extends` just in case. The actual configuration
for this project itself still lives at `src/tsconfig.json`, but this prevents
issues when anything (like the `ts-node` that gets loaded in `commitlint`
from trying to read a non-existent `dist/tsconfig.json` if it does read
the `tsconfig.json` at the root of the project.