fix(scripts/commit-msg): prevent commitlint from ever type-checking configuration
**@commitlint/cli** added support for TypeScript-based configurations via `ts-node`, and that has caused a good deal of grief:

https://github.com/conventional-changelog/commitlint/issues/3420
https://github.com/conventional-changelog/commitlint/pull/3351
https://github.com/conventional-changelog/commitlint/issues/3218
https://github.com/conventional-changelog/commitlint/issues/3256

https://github.com/hoverinc/web-react/pull/1792

Both because it's a heavy-ass dependency and because it requires configuration to
work properly. It seems like project-local configuration can affect it's invocation
of ts-node, which happens even JavaScript configuration files now?... lame.

Anyways, this ensures we avoid any type-checking that happens in it's ts-node
invocation until we get a better solution upstream (or contribute one). Ideally,
installing ts-node would be opt-in etc.